move OsmData from osm_maps_functions.py to osm_data.py
- pylint back at 10.00/10
